+++ python/src/pypalmsens/__init__.py
@@ -27,8 +27,8 @@ from . import (
     stages,
     types,
 )
-from ._instruments.comm import CommProtocol
-from ._instruments.comm_async import CommProtocolAsync
+from ._instruments.comm_protocol import CommProtocol
+from ._instruments.comm_protocol_async import CommProtocolAsync
 from ._instruments.filesystem import DeviceFileSystem, DevicePath
 from ._instruments.filesystem_async import DeviceFileSystemAsync
 from ._instruments.instrument import Instrument, discover, discover_async
