+++ python/src/pypalmsens/_instruments/comm_protocol.py
